refactor: JdbcTemplate와 selectJdbcTemplate 병합
