+++ app/src/main/java/com/techcourse/dao/JdbcTemplate.java
@@ -1,8 +1,8 @@
 package com.techcourse.dao;
 
-import com.techcourse.domain.User;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import javax.sql.DataSource;
 import org.slf4j.Logger;
@@ -11,9 +11,6 @@ import org.slf4j.LoggerFactory;
 public abstract class JdbcTemplate {
 
     private static final Logger LOG = LoggerFactory.getLogger(JdbcTemplate.class);
-
-    protected JdbcTemplate() {
-    }
 
     protected abstract String createQuery();
 
@@ -33,7 +30,39 @@ public abstract class JdbcTemplate {
         }
     }
 
-    protected abstract DataSource getDataSource();
-
     protected abstract void setValues(PreparedStatement pstmt) throws SQLException;
+
+    private ResultSet executeQuery(PreparedStatement pstmt) throws SQLException {
+        return pstmt.executeQuery();
+    }
+
+    public Object query() {
+        final String sql = createQuery();
+
+        ResultSet resultSet = null;
+        try (Connection conn = getDataSource().getConnection();
+            PreparedStatement pstmt = conn.prepareStatement(sql)) {
+
+            setValues(pstmt);
+
+            LOG.debug("query : {}", sql);
+            resultSet = executeQuery(pstmt);
+            return mapRow(resultSet);
+
+        } catch (SQLException e) {
+            LOG.error(e.getMessage(), e);
+            throw new RuntimeException(e);
+        } finally {
+            try {
+                if (resultSet != null) {
+                    resultSet.close();
+                }
+            } catch (SQLException ignored) {
+            }
+        }
+    }
+
+    protected abstract Object mapRow(ResultSet resultSet) throws SQLException;
+
+    protected abstract DataSource getDataSource();
 }
