+++ docs/README.md
@@ -4,7 +4,19 @@
 - [x] 콜백 패턴 공부
 - [x] 리플렉션 테스트코드 완성
 - [x] userDao 기능 구현
+
 - [ ] jdbcTemplate 구현
+  - [x] userDao와 insert, update 분리하기
+  - [x] user와 jdbcTemplate 의존성 분리하기
+  - [ ] userDao와 select 분리하기
+  - [ ] selectJdbcTemplate와 jdbcTemplate 병합하기
+  - [ ] mapRow 인터페이스 분리
+  - [ ] setValues 인터페이스 분리
+  - [ ] 라이브러리 확장하기
+    - [ ] 커스텀 익셉션 추가
+    - [ ] Object에서 제너릭으로 개선하기
+    - [ ] preparedStatementSetter를 가변인자로 변경
+    - [ ] 람다식을 이용하여 코드량 줄이기
 - [ ] DI 구현
 - [ ] controller service Repository 분리
 - [ ] pages enum 구현 후 분리
